WasteLine form conditionals (#582)
* adjust rollup chunkSize WarningLimit to 1000 kb

* convert dotHazardous field to controlled input

* convert epaWaste to controlled field

We need to be able to read the state of the field 'epaWaste' in order to
conditionally display/disable other fields.

* add zod validation schemas for dotInformation field and add as nested field in WasteLine

while the dotInformation is optional if the shipement is not dotHazardous, we've also added a refine method to the wasteline schema so if dotHazardous is true, both fields in dotInformation is required

* add static dotIdNumbers array temporaryily

* add error message when DOT ID number is not provided

* simplify wasteline zod refinements, remove old unused wasteline interfaces

* add wasteDescription validation

* add refinement that waste description is required if the waste is not DOT hazardous material

* update logic for dotHazardous and epaWaste switches

here's our logic, if dotHazardous then epaWaste can be true or false. However, if epaWaste is True then dotHazardous must be true

* add callback function to toggle epaWaste and dotHazardous with unit test

epaWaste and dotHazardous field depend on each other if epaWaste is true, then dotHazardous must be true. Converse, if dotHazardous is false, epaWaste must be false

* add integration test showing that when epaWaste is false, federal waste code select is disabled

* when epaWaste is set to false, all federal waste codes are removed from the wasteline